GEOMESA-1076 Move sft and config resolution classes
